refactor: split variables.js into hash-ops, list-ops, and variables
Extract hash operation helpers to variable-hash-ops.js and list
operation handlers to variable-list-ops.js from the 1134-line
variables.js.

Refs #357
