+++ packages/scratch-gui/src/lib/ruby-to-blocks-converter/variables.js
@@ -1,6 +1,7 @@
 import {defineMessages} from 'react-intl';
-import _ from 'lodash';
 import {RubyToBlocksConverterError} from './errors';
+import {convertToListBlock, adjustIndex, convertHashGet, convertHashSet} from './variable-hash-ops';
+import {registerListOperations} from './variable-list-ops';
 
 const messages = defineMessages({
     arraySyntaxNotAvailableInV1: {
@@ -35,631 +36,9 @@ const messages = defineMessages({
  */
 const VariablesConverter = {
     register: function (converter) {
-        /**
-         * Convert a data_variable block to data_listcontents for list operations.
-         * When $a (global) or `@a` (instance) is used with array methods like .push(),
-         * the variable read creates a data_variable block, but list operations need
-         * a data_listcontents block with LIST field.
-         * @param {object} block - The receiver block to convert
-         * @returns {{block: object|null, converted: boolean}} The converted block and whether conversion happened
-         */
-        const convertToListBlock = function (block) {
-            if (!converter._isBlock(block)) return {block: null, converted: false};
-
-            // Already a list block, no conversion needed
-            if (converter.isListBlock(block)) return {block, converted: false};
-
-            if (block.opcode !== 'data_variable') return {block: null, converted: false};
-
-            // Only convert data_variable blocks in version 2
-            if (converter.version < 2) {
-                throw new RubyToBlocksConverterError(
-                    converter._context.currentNode,
-                    converter._translator(messages.arraySyntaxNotAvailableInV1)
-                );
-            }
-
-            const varName = block.fields.VARIABLE.value;
-            const variable = converter._context.variables[varName] ||
-                converter._context.localVariables[varName];
-            if (!variable) return {block: null, converted: false};
-
-            let prefixedName;
-            if (variable.scope === 'global') {
-                prefixedName = `$${varName}`;
-            } else if (variable.scope === 'instance') {
-                prefixedName = `@${varName}`;
-            } else if (variable.scope === 'local') {
-                prefixedName = variable.originalName;
-            } else {
-                return {block: null, converted: false};
-            }
-
-            const listVar = converter._lookupOrCreateList(prefixedName);
-
-            // Convert the block in-place
-            block.opcode = 'data_listcontents';
-            delete block.fields.VARIABLE;
-            block.fields.LIST = {
-                name: 'LIST',
-                id: listVar.id,
-                value: listVar.name,
-                variableType: listVar.type
-            };
-
-            return {block, converted: true};
-        };
-
-        /**
-         * Adjust a 0-indexed Ruby array index to 1-indexed Scratch list index.
-         * Always wraps in operator_add(index, 1) with `@ruby`:array:index comment
-         * to enable round-trip conversion.
-         * @param {object} index - The index block or value.
-         * @param {boolean} converted - Whether the receiver was converted to a list block.
-         */
-        const adjustIndex = function (index, converted) {
-            if (!converted) return index;
-            const addBlock = converter._createBlock('operator_add', 'value');
-            converter._addNumberInput(addBlock, 'NUM1', 'math_number', index, 0);
-            converter._addNumberInput(addBlock, 'NUM2', 'math_number', 1, 0);
-            addBlock.comment = converter._createComment('@ruby:array:index', addBlock.id);
-            return addBlock;
-        };
-
-        /**
-         * Convert a hash bracket read: $a[:key] or $a["key"].
-         * Generates data_itemoflist(INDEX: data_itemnumoflist(key, keys_list), values_list).
-         * @param {object} receiver - The receiver block (data_variable).
-         * @param {*} keyArg - The key argument (symbol or string).
-         * @returns {object|null} The generated block, or null on failure.
-         */
-        const convertHashGet = function (receiver, keyArg) {
-            if (!converter._isBlock(receiver) || receiver.opcode !== 'data_variable') return null;
-
-            const varName = receiver.fields.VARIABLE.value;
-            const variable = converter._context.variables[varName] ||
-                converter._context.localVariables[varName];
-            if (!variable) return null;
-
-            let prefixedName;
-            if (variable.scope === 'global') {
-                prefixedName = `$${varName}`;
-            } else if (variable.scope === 'instance') {
-                prefixedName = `@${varName}`;
-            } else if (variable.scope === 'local') {
-                prefixedName = variable.originalName;
-            } else {
-                return null;
-            }
-
-            const keysListName = converter._hashKeysListName(prefixedName);
-            const valuesListName = converter._hashValuesListName(prefixedName);
-            const keysList = converter._lookupOrCreateList(keysListName);
-            const valuesList = converter._lookupOrCreateList(valuesListName);
-
-            let keyStr;
-            let commentMarker;
-            if (converter._isSymbol(keyArg)) {
-                const symName = converter._getSymbolValue(keyArg);
-                keyStr = `:${symName}`;
-                commentMarker = '@ruby:hash:get:sym';
-            } else {
-                keyStr = converter._isPrimitive(keyArg) ? keyArg.value : keyArg;
-                commentMarker = '@ruby:hash:get:str';
-            }
-
-            // Create data_itemnumoflist block for key lookup
-            const numBlock = converter._createBlock('data_itemnumoflist', 'value', {
-                fields: {
-                    LIST: {
-                        name: 'LIST',
-                        id: keysList.id,
-                        value: keysList.name,
-                        variableType: keysList.type
-                    }
-                }
-            });
-            converter._addTextInput(numBlock, 'ITEM', keyStr, 'thing');
-
-            // Create data_itemoflist block for value retrieval
-            const block = converter._changeBlock(receiver, 'data_itemoflist', 'value');
-            delete block.fields.VARIABLE;
-            block.fields.LIST = {
-                name: 'LIST',
-                id: valuesList.id,
-                value: valuesList.name,
-                variableType: valuesList.type
-            };
-            converter._addNumberInput(block, 'INDEX', 'math_integer', numBlock, 1);
-
-            // Include @ruby:lvar prefix for local variables
-            if (variable.scope === 'local') {
-                commentMarker =
-                    `@ruby:lvar:${variable.originalName}:${variable.scopeIndex},${commentMarker}`;
-            }
-            block.comment = converter._createComment(commentMarker, block.id);
-
-            return block;
-        };
-
-        /**
-         * Convert a hash bracket write (upsert): $a[:key] = value.
-         * Generates delete+push pattern (4 blocks).
-         * @param {object} receiver - The receiver block (data_variable).
-         * @param {*} keyArg - The key argument (symbol or string).
-         * @param {*} valueArg - The value to set.
-         * @returns {object|null} The generated block chain, or null on failure.
-         */
-        const convertHashSet = function (receiver, keyArg, valueArg) {
-            if (!converter._isBlock(receiver) || receiver.opcode !== 'data_variable') return null;
-
-            const varName = receiver.fields.VARIABLE.value;
-            const variable = converter._context.variables[varName] ||
-                converter._context.localVariables[varName];
-            if (!variable) return null;
-
-            let prefixedName;
-            if (variable.scope === 'global') {
-                prefixedName = `$${varName}`;
-            } else if (variable.scope === 'instance') {
-                prefixedName = `@${varName}`;
-            } else if (variable.scope === 'local') {
-                prefixedName = variable.originalName;
-            } else {
-                return null;
-            }
-
-            const keysListName = converter._hashKeysListName(prefixedName);
-            const valuesListName = converter._hashValuesListName(prefixedName);
-            const keysList = converter._lookupOrCreateList(keysListName);
-            const valuesList = converter._lookupOrCreateList(valuesListName);
-
-            let keyStr;
-            let commentMarker;
-            if (converter._isSymbol(keyArg)) {
-                const symName = converter._getSymbolValue(keyArg);
-                keyStr = `:${symName}`;
-                commentMarker = '@ruby:hash:set:sym';
-            } else {
-                keyStr = converter._isPrimitive(keyArg) ? keyArg.value : keyArg;
-                commentMarker = '@ruby:hash:set:str';
-            }
-
-            // Handle symbol values
-            let valueItem;
-            if (converter._isPrimitive(valueArg) && valueArg.type === 'sym') {
-                valueItem = converter._symbolToBlock(valueArg.value, valueArg.node);
-            } else if (converter._isNumber(valueArg)) {
-                valueItem = converter._isPrimitive(valueArg) ?
-                    valueArg.value.toString() : valueArg.toString();
-            } else if (converter._isString(valueArg)) {
-                valueItem = converter._isPrimitive(valueArg) ? valueArg.value : valueArg;
-            } else {
-                valueItem = valueArg;
-            }
-
-            // Block 1: delete from values list
-            const deleteValuesBlock = converter._changeBlock(receiver, 'data_deleteoflist', 'statement');
-            delete deleteValuesBlock.fields.VARIABLE;
-            deleteValuesBlock.fields.LIST = {
-                name: 'LIST',
-                id: valuesList.id,
-                value: valuesList.name,
-                variableType: valuesList.type
-            };
-            const numBlock1 = converter._createBlock('data_itemnumoflist', 'value', {
-                fields: {
-                    LIST: {
-                        name: 'LIST',
-                        id: keysList.id,
-                        value: keysList.name,
-                        variableType: keysList.type
-                    }
-                }
-            });
-            converter._addTextInput(numBlock1, 'ITEM', keyStr, 'thing');
-            converter._addNumberInput(deleteValuesBlock, 'INDEX', 'math_integer', numBlock1, 1);
-
-            // Include @ruby:lvar prefix for local variables
-            if (variable.scope === 'local') {
-                commentMarker =
-                    `@ruby:lvar:${variable.originalName}:${variable.scopeIndex},${commentMarker}`;
-            }
-            deleteValuesBlock.comment = converter._createComment(commentMarker, deleteValuesBlock.id);
-
-            // Block 2: delete from keys list
-            const deleteKeysBlock = converter._createBlock('data_deleteoflist', 'statement', {
-                fields: {
-                    LIST: {
-                        name: 'LIST',
-                        id: keysList.id,
-                        value: keysList.name,
-                        variableType: keysList.type
-                    }
-                }
-            });
-            const numBlock2 = converter._createBlock('data_itemnumoflist', 'value', {
-                fields: {
-                    LIST: {
-                        name: 'LIST',
-                        id: keysList.id,
-                        value: keysList.name,
-                        variableType: keysList.type
-                    }
-                }
-            });
-            converter._addTextInput(numBlock2, 'ITEM', keyStr, 'thing');
-            converter._addNumberInput(deleteKeysBlock, 'INDEX', 'math_integer', numBlock2, 1);
-            deleteKeysBlock.comment = converter._createComment(
-                '@ruby:hash:set:delete:key', deleteKeysBlock.id
-            );
-
-            // Block 3: push key
-            const pushKeyBlock = converter._createBlock('data_addtolist', 'statement', {
-                fields: {
-                    LIST: {
-                        name: 'LIST',
-                        id: keysList.id,
-                        value: keysList.name,
-                        variableType: keysList.type
-                    }
-                }
-            });
-            converter._addTextInput(pushKeyBlock, 'ITEM', keyStr, 'thing');
-            pushKeyBlock.comment = converter._createComment(
-                '@ruby:hash:set:push:key', pushKeyBlock.id
-            );
-
-            // Block 4: push value
-            const pushValueBlock = converter._createBlock('data_addtolist', 'statement', {
-                fields: {
-                    LIST: {
-                        name: 'LIST',
-                        id: valuesList.id,
-                        value: valuesList.name,
-                        variableType: valuesList.type
-                    }
-                }
-            });
-            converter._addTextInput(
-                pushValueBlock, 'ITEM',
-                converter._isNumber(valueItem) ? valueItem.toString() : valueItem, 'thing'
-            );
-            pushValueBlock.comment = converter._createComment(
-                '@ruby:hash:set:push:value', pushValueBlock.id
-            );
-
-            return converter._linkBlocks([
-                deleteValuesBlock, deleteKeysBlock, pushKeyBlock, pushValueBlock
-            ]);
-        };
-
-        converter.registerOnSend('self', 'show_variable', 1, params => {
-            const {args} = params;
-            if (!converter._isString(args[0])) return null;
-
-            const variable = converter._lookupOrCreateVariable(args[0]);
-            if (variable.scope === 'global' || variable.scope === 'instance') {
-                return converter._createBlock('data_showvariable', 'statement', {
-                    fields: {
-                        VARIABLE: {
-                            name: 'VARIABLE',
-                            id: variable.id,
-                            value: variable.name,
-                            variableType: variable.type
-                        }
-                    }
-                });
-            }
-            return null;
-        });
-
-        converter.registerOnSend('self', 'hide_variable', 1, params => {
-            const {args} = params;
-            if (!converter._isString(args[0])) return null;
-
-            const variable = converter._lookupOrCreateVariable(args[0]);
-            if (variable.scope === 'global' || variable.scope === 'instance') {
-                return converter._createBlock('data_hidevariable', 'statement', {
-                    fields: {
-                        VARIABLE: {
-                            name: 'VARIABLE',
-                            id: variable.id,
-                            value: variable.name,
-                            variableType: variable.type
-                        }
-                    }
-                });
-            }
-            return null;
-        });
-
-        converter.registerOnSend('self', 'list', 1, params => {
-            const {args} = params;
-            if (!converter._isString(args[0])) return null;
-
-            if (converter.version >= 2) {
-                throw new RubyToBlocksConverterError(
-                    params.node,
-                    converter._translator(messages.listSyntaxNotAvailableInV2)
-                );
-            }
-
-            const variable = converter._lookupOrCreateList(args[0]);
-            if (variable.scope === 'global' || variable.scope === 'instance') {
-                return converter._createBlock('data_listcontents', 'value_variable', {
-                    fields: {
-                        LIST: {
-                            name: 'LIST',
-                            id: variable.id,
-                            value: variable.name,
-                            variableType: variable.type
-                        }
-                    }
-                });
-            }
-            return null;
-        });
-
-        converter.registerOnSend('self', 'show_list', 1, params => {
-            const {args} = params;
-            if (!converter._isString(args[0])) return null;
-
-            const variable = converter._lookupOrCreateList(args[0]);
-            if (variable.scope === 'global' || variable.scope === 'instance') {
-                return converter._createBlock('data_showlist', 'statement', {
-                    fields: {
-                        LIST: {
-                            name: 'LIST',
-                            id: variable.id,
-                            value: variable.name,
-                            variableType: variable.type
-                        }
-                    }
-                });
-            }
-            return null;
-        });
-
-        converter.registerOnSend('self', 'hide_list', 1, params => {
-            const {args} = params;
-            if (!converter._isString(args[0])) return null;
-
-            const variable = converter._lookupOrCreateList(args[0]);
-            if (variable.scope === 'global' || variable.scope === 'instance') {
-                return converter._createBlock('data_hidelist', 'statement', {
-                    fields: {
-                        LIST: {
-                            name: 'LIST',
-                            id: variable.id,
-                            value: variable.name,
-                            variableType: variable.type
-                        }
-                    }
-                });
-            }
-            return null;
-        });
-
-        converter.registerOnSend('variable', 'push', 1, params => {
-            const {receiver, args} = params;
-            if (!converter._isStringOrBlock(args[0]) && !converter._isNumberOrBlock(args[0])) return null;
-
-            const {block: listBlock, converted} = convertToListBlock(receiver);
-            const recv = converted ? listBlock : receiver;
-            if (!recv) return null;
-
-            const block = converter._changeBlock(recv, 'data_addtolist', 'statement');
-            converter._addTextInput(
-                block, 'ITEM', converter._isNumber(args[0]) ? args[0].toString() : args[0], 'thing'
-            );
-            return block;
-        });
-
-        converter.registerOnSend('variable', '<<', 1, params => {
-            const {receiver, args} = params;
-            if (!converter._isStringOrBlock(args[0]) && !converter._isNumberOrBlock(args[0])) return null;
-
-            // convertToListBlock will throw error in v1 if receiver is data_variable
-            const {block: listBlock, converted} = convertToListBlock(receiver);
-            const recv = converted ? listBlock : receiver;
-            if (!recv) return null;
-
-            const block = converter._changeBlock(recv, 'data_addtolist', 'statement');
-            converter._addTextInput(
-                block, 'ITEM', converter._isNumber(args[0]) ? args[0].toString() : args[0], 'thing'
-            );
-            return block;
-        });
-
-        converter.registerOnSend('variable', 'delete_at', 1, params => {
-            const {receiver, args} = params;
-            if (!converter._isNumberOrBlock(args[0])) return null;
-
-            const {block: listBlock, converted} = convertToListBlock(receiver);
-            const recv = converted ? listBlock : receiver;
-            if (!recv) return null;
-            const index = adjustIndex(args[0], converted);
-
-            const block = converter._changeBlock(recv, 'data_deleteoflist', 'statement');
-            converter._addNumberInput(block, 'INDEX', 'math_integer', index, 1);
-            return block;
-        });
-
-        converter.registerOnSend('variable', 'clear', 0, params => {
-            const {receiver} = params;
-
-            const {block: listBlock, converted} = convertToListBlock(receiver);
-            const recv = converted ? listBlock : receiver;
-            if (!recv) return null;
-
-            return converter._changeBlock(recv, 'data_deletealloflist', 'statement');
-        });
-
-        converter.registerOnSend('variable', 'insert', 2, params => {
-            const {receiver, args} = params;
-            if (!converter._isNumberOrBlock(args[0])) return null;
-            if (!converter._isStringOrBlock(args[1]) && !converter._isNumberOrBlock(args[1])) return null;
-
-            const {block: listBlock, converted} = convertToListBlock(receiver);
-            const recv = converted ? listBlock : receiver;
-            if (!recv) return null;
-            const index = adjustIndex(args[0], converted);
-
-            const block = converter._changeBlock(recv, 'data_insertatlist', 'statement');
-            converter._addNumberInput(block, 'INDEX', 'math_integer', index, 1);
-            converter._addTextInput(
-                block, 'ITEM', converter._isNumber(args[1]) ? args[1].toString() : args[1], 'thing'
-            );
-            return block;
-        });
-
-        converter.registerOnSend('variable', '[]=', 2, params => {
-            const {receiver, args} = params;
-
-            // Hash write: $a[:key] = value or $a["key"] = value
-            if (converter._isSymbol(args[0]) || converter._isString(args[0])) {
-                if (converter.version < 2) {
-                    throw new RubyToBlocksConverterError(
-                        converter._context.currentNode,
-                        converter._translator(messages.hashSyntaxNotAvailableInV1)
-                    );
-                }
-                if (!converter._isStringOrBlock(args[1]) && !converter._isNumberOrBlock(args[1]) &&
-                    !(converter._isPrimitive(args[1]) && args[1].type === 'sym')) return null;
-                return convertHashSet(receiver, args[0], args[1]);
-            }
-
-            if (!converter._isNumberOrBlock(args[0])) return null;
-            if (!converter._isStringOrBlock(args[1]) && !converter._isNumberOrBlock(args[1])) return null;
-
-            const {block: listBlock, converted} = convertToListBlock(receiver);
-            const recv = converted ? listBlock : receiver;
-            if (!recv) return null;
-            const index = adjustIndex(args[0], converted);
-
-            const block = converter._changeBlock(recv, 'data_replaceitemoflist', 'statement');
-            converter._addNumberInput(block, 'INDEX', 'math_integer', index, 1);
-            converter._addTextInput(
-                block, 'ITEM', converter._isNumber(args[1]) ? args[1].toString() : args[1], 'thing'
-            );
-            return block;
-        });
-
-        converter.registerOnSend('variable', '[]', 1, params => {
-            const {receiver, args} = params;
-
-            // Hash access: $a[:key] or $a["key"]
-            if (converter._isSymbol(args[0]) || converter._isString(args[0])) {
-                if (converter.version < 2) {
-                    throw new RubyToBlocksConverterError(
-                        converter._context.currentNode,
-                        converter._translator(messages.hashSyntaxNotAvailableInV1)
-                    );
-                }
-                return convertHashGet(receiver, args[0]);
-            }
-
-            if (!converter._isNumberOrBlock(args[0])) return null;
-
-            const {block: listBlock, converted} = convertToListBlock(receiver);
-            if (converted && listBlock) {
-                const index = adjustIndex(args[0], true);
-                const block = converter._changeBlock(listBlock, 'data_itemoflist', 'value');
-                converter._addNumberInput(block, 'INDEX', 'math_integer', index, 1);
-                return block;
-            }
-
-            if (converter._isBlock(receiver) && converter.isListBlock(receiver)) {
-                const block = converter._changeBlock(receiver, 'data_itemoflist', 'value');
-                converter._addNumberInput(block, 'INDEX', 'math_integer', args[0], 1);
-                return block;
-            }
-
-            return null;
-        });
-
-        converter.registerOnSend('variable', 'index', 1, params => {
-            const {receiver, args} = params;
-            if (!converter._isStringOrBlock(args[0]) && !converter._isNumberOrBlock(args[0])) return null;
-
-            const {block: listBlock, converted} = convertToListBlock(receiver);
-            const recv = converted ? listBlock : receiver;
-            if (!recv) return null;
-
-            const block = converter._changeBlock(recv, 'data_itemnumoflist', 'value');
-            converter._addTextInput(
-                block, 'ITEM', converter._isNumber(args[0]) ? args[0].toString() : args[0], 'thing'
-            );
-
-            // Wrap in operator_subtract(result, 1) for 0-indexed return value
-            if (converted) {
-                const subtractBlock = converter._createBlock('operator_subtract', 'value');
-                converter._addInput(subtractBlock, 'NUM1', block);
-                converter._addNumberInput(subtractBlock, 'NUM2', 'math_number', 1, 0);
-                subtractBlock.comment = converter._createComment(
-                    '@ruby:array:index', subtractBlock.id
-                );
-                return subtractBlock;
-            }
-
-            return block;
-        });
-
-        converter.registerOnSend('variable', 'length', 0, params => {
-            const {receiver} = params;
-
-            const {block: listBlock, converted} = convertToListBlock(receiver);
-            if (converted && listBlock) {
-                return converter._changeBlock(listBlock, 'data_lengthoflist', 'value');
-            }
-
-            if (converter._isBlock(receiver) && converter.isListBlock(receiver)) {
-                return converter._changeBlock(receiver, 'data_lengthoflist', 'value');
-            }
-            return null;
-        });
-
-        converter.registerOnSend('variable', 'include?', 1, params => {
-            const {receiver, args} = params;
-            if (!converter._isStringOrBlock(args[0]) && !converter._isNumberOrBlock(args[0])) return null;
-
-            let recv = receiver;
-            if (converter.version >= 2) {
-                const {block: listBlock, converted} = convertToListBlock(receiver);
-                recv = converted ? listBlock : receiver;
-                if (!recv) return null;
-            }
-
-            const block = converter._changeBlock(recv, 'data_listcontainsitem', 'value_boolean');
-            converter._addTextInput(
-                block, 'ITEM', converter._isNumber(args[0]) ? args[0].toString() : args[0], 'thing'
-            );
-            return block;
-        });
-
-        converter.registerOnSend('variable', 'empty?', 0, params => {
-            if (converter.version < 2) return null;
-
-            const {receiver} = params;
-
-            const {block: listBlock, converted} = convertToListBlock(receiver);
-            if (!converted || !listBlock) return null;
-
-            const name = 'empty?';
-            const index = (converter._context.methodCallIndices[name] || 0) + 1;
-            converter._context.methodCallIndices[name] = index;
-            const commentText = `@ruby:method:${name}:${index}`;
-
-            const lengthBlock = converter._changeBlock(listBlock, 'data_lengthoflist', 'value');
-            lengthBlock.comment = converter._createComment(commentText, lengthBlock.id);
-
-            const block = converter._createBlock('operator_equals', 'value_boolean');
-            converter._addInput(block, 'OPERAND1', lengthBlock, converter._createTextBlock(''));
-            converter._addTextInput(block, 'OPERAND2', '0', '50');
-            block.comment = converter._createComment(commentText, block.id);
-            return block;
-        });
+        registerListOperations(
+            converter, messages, convertToListBlock, adjustIndex, convertHashGet, convertHashSet
+        );
 
         // Operator to opcode mapping for compound assignments
         const COMPOUND_OPERATOR_MAP = {
